Move `FileId` to the `node` module
