+++ crates/apollo-compiler/src/database/mod.rs
@@ -7,6 +7,5 @@ mod sources;
 pub(crate) use db::RootDatabase;
 pub(crate) use inputs::{InputDatabase, InputStorage};
 pub(crate) use repr::{ReprDatabase, ReprStorage};
-pub use sources::FileId;
 pub(crate) use sources::Source;
 pub(crate) use sources::SourceType;
