+++ crates/apollo-compiler/src/parser.rs
@@ -6,10 +6,10 @@ use crate::schema::SchemaBuilder;
 use crate::validation::Details;
 use crate::validation::DiagnosticList;
 use crate::validation::FileId;
-use crate::validation::NodeLocation;
 use crate::validation::Valid;
 use crate::validation::WithErrors;
 use crate::ExecutableDocument;
+use crate::NodeLocation;
 use crate::Schema;
 use indexmap::IndexMap;
 use std::path::Path;
